requirements.txt
APScheduler==3.10.1
cairocffi==1.6.0
CairoSVG==2.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cssselect2==0.7.0
defusedxml==0.7.1
idna==3.4
MLB-StatsAPI==1.6.1
Pillow==10.0.0
pycparser==2.21
pytz==2023.3
requests==2.31.0
six==1.16.0
tinycss2==1.2.1
tzlocal==5.0.1
urllib3==2.0.3
webencodings==0.5.1
